feat: remove date picker button single
